# YJIT: Instrument global allocations on stats build