For stats build, this PR allows you to measure how much memory YJIT uses in Rust.

Enabling YJIT currently consumes a lot more memory than what's used for code pages. I think it's nice to monitor changes in allocated bytes before attempting to reduce it because some changes could come at the cost of implementation complexity and we shouldn't do it when it's not really effective.

## Example
The current state of 25 itrs on railsbench:

```
inline_code_size:         4855668
outlined_code_size:       4852936
freed_code_size:                0
yjit_alloc_size:         25114317
```

When 9.7MB is used for JIT code, 25.1MB is used for Rust's global allocations.

## Benchmark
Its performance impact on stats build seems negligible. Here's the yjit-bench result on railsbench.

### Before
```
itr #1: 1503ms
itr #2: 1164ms
itr #3: 1148ms
itr #4: 1179ms
itr #5: 1114ms
itr #6: 1173ms
itr #7: 1165ms
itr #8: 1153ms
itr #9: 1173ms
itr #10: 1152ms
itr #11: 1149ms
itr #12: 1180ms
itr #13: 1166ms
itr #14: 1138ms
itr #15: 1179ms
itr #16: 1149ms
itr #17: 1163ms
itr #18: 1173ms
itr #19: 1144ms
itr #20: 1166ms
itr #21: 1171ms
itr #22: 1169ms
itr #23: 1149ms
itr #24: 1146ms
itr #25: 1174ms
```

### After
```
itr #1: 1443ms
itr #2: 1135ms
itr #3: 1131ms
itr #4: 1149ms
itr #5: 1143ms
itr #6: 1130ms
itr #7: 1136ms
itr #8: 1150ms
itr #9: 1164ms
itr #10: 1134ms
itr #11: 1134ms
itr #12: 1178ms
itr #13: 1127ms
itr #14: 1134ms
itr #15: 1109ms
itr #16: 1110ms
itr #17: 1155ms
itr #18: 1140ms
itr #19: 1130ms
itr #20: 1150ms
itr #21: 1163ms
itr #22: 1155ms
itr #23: 1144ms
itr #24: 1165ms
itr #25: 1130ms
```