@k0kubun 2022-11-11
Indeed https://github.com/ruby/ruby/pull/6712/commits/9c3c0b3209a9a97107ed387e4fa5fae0d73f05c8. Thank you for the finding. This seems to return the same number.